s5e9945: libaudioproxy: Raise stop thresholds
Using ULONG_MAX instead of UINT_MAX ensures that the affected stop
thresholds are never smaller than their stream's boundaries and
snd_pcm_playback_data continues to evaluate as true when compiling for
aarch64, this resolves the following error in particular:

  E audio_hw_proxy: proxy-voice_rx_start: Voice Call RX PCM Device(/dev/snd/pcmC0D104p) with SR(48000) PF(0) CC(2) cannot be started as error(cannot start channel: Broken pipe)

+++ libaudioproxy/audio_usb_proxy.c
@@ -930,7 +930,7 @@ static void usb_open_out_proxy(struct audio_proxy_usb *aproxy_usb)
             /* Update period-size using updated config rate */
             ppcmconfig->period_count = DEFAULT_USB_PERIOD_COUNT;
             ppcmconfig->period_size = (ppcmconfig->rate * DEFAULT_USB_PLAYBACK_DURATION) / 1000;
-            ppcmconfig->stop_threshold = UINT_MAX;
+            ppcmconfig->stop_threshold = ULONG_MAX;
             aproxy_usb->usb_out_pcm = pcm_open(aproxy_usb->usb_out_pcm_card,
                                                 aproxy_usb->usb_out_pcm_device,
                                                 flags, ppcmconfig);
